swaps mobile/desktop at breakpoints
