+++ package.json
@@ -14,7 +14,8 @@
     "prettier": "^3.3.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-swipeable": "^7.0.1"
+    "react-swipeable": "^7.0.1",
+    "use-breakpoint": "^4.0.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
