+++ src/AppMobile.tsx
@@ -7,15 +7,14 @@ import {
   hardDropBlock,
   holdAndPopHeld,
   rotateBlock,
+  setAllowedInput,
   shiftBlock,
   startGame,
   tickGameClock,
 } from './Tetris';
-import ThemeSong from './assets/ThemeSong.mp3';
 import { CONFIG } from './TetrisConfig';
 import PieceDisplay from './components/PieceDisplay';
 import { SwipeEventData, useSwipeable } from 'react-swipeable';
-const music = new Audio(ThemeSong);
 
 const cellBorderStyles = ['outset', 'none'];
 
@@ -44,32 +43,58 @@ export default function MobileApp() {
       document.body.style.overflow = 'scroll';
     };
   }, []);
+  const softDrop = () => {
+    setGameState((prevState) =>
+      setAllowedInput(shiftBlock(prevState, 'D'), 'shift', false)
+    );
+    setTimeout(
+      () =>
+        setGameState((prevState) => setAllowedInput(prevState, 'shift', true)),
+      CONFIG.POLL_RATES.shift
+    );
+  };
 
   const [lastShiftSteps, setLastShiftSteps] = useState(0); //how many steps from swipe origin we have shifted the current block
+  const [swipeStart, setSwipeStart] = useState(0);
   const PIXELS_PER_SHIFT_STEP = 32; //how many pixels of swiping corresponds to a side shift step
   const handleSwipeStart = () => {
+    setLastShiftSteps(0);
+    setSwipeStart(Date.now());
+  };
+  const handleSwipeEnd = () => {
+    setSwipeStart(0);
     setLastShiftSteps(0);
   };
   const handleSwiping = (e: SwipeEventData) => {
     const dx = e.deltaX;
+    const dy = e.deltaY;
     const shiftSteps = Math.floor(dx / PIXELS_PER_SHIFT_STEP);
     if (shiftSteps !== lastShiftSteps) {
       const direction: Direction = shiftSteps > lastShiftSteps ? 'R' : 'L';
       setGameState((prevState) => shiftBlock(prevState, direction));
       setLastShiftSteps(shiftSteps);
     }
+    //todo: soft drop when swiping down?
+    if (e.absY > 100) {
+      softDrop();
+    }
   };
   const handleTap = () => {
     setGameState((prevState) => rotateBlock(prevState, 'CW'));
   };
   const handleSwipeDown = () => {
+    const swipeDuration = Date.now() - swipeStart;
+    if (swipeDuration > 500) return;
     setGameState((prevState) => hardDropBlock(prevState));
   };
   const handleSwipeUp = () => {
+    const swipeDuration = Date.now() - swipeStart;
+    if (swipeDuration > 500) return;
     setGameState((prevState) => holdAndPopHeld(prevState));
   };
 
   const handlers = useSwipeable({
+    onSwiped: handleSwipeEnd,
     onSwiping: handleSwiping,
     onSwipeStart: handleSwipeStart,
     onTap: handleTap,
@@ -111,11 +136,18 @@ export default function MobileApp() {
       </div>
       {/* Game over modal */}
       {gameState.over && (
-        <div className="absolute left-1/2 top-1/2 h-full w-full -translate-x-1/2 -translate-y-1/2 bg-slate-900 bg-opacity-80">
-          <div className="text-default flex h-full w-full animate-flash flex-col items-center justify-center gap-5 px-8 py-4 text-5xl">
+        <div className="absolute left-1/2 top-1/2 flex h-full w-full -translate-x-1/2 -translate-y-1/2 flex-col items-center justify-center bg-slate-900 bg-opacity-80">
+          <div className="text-default flex w-full animate-flash flex-col items-center justify-center gap-5 px-8 py-4 text-5xl">
             <span className="">GAME OVER</span>
             <span>Score: {gameState.score}</span>
           </div>
+          <button
+            onClick={() => setGameState(startGame(gameState))}
+            style={{ borderStyle: 'inset' }}
+            className="text-default mt-4 border-[8px] border-green-500 bg-slate-900 bg-opacity-80 px-4 py-4"
+          >
+            Restart
+          </button>
         </div>
       )}
       {/* Startgame modal */}
