+++ src/main.tsx
@@ -1,12 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App.tsx';
 import './index.css';
-import MobileApp from './AppMobile.tsx';
-
+import App from './App.tsx';
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <MobileApp />
-    {/* <App /> */}
+    <App />
   </React.StrictMode>
 );
